Added percentile histograms (issue #72)
* Fixed LogbackMetrics to not count logging events whose levels are below the effective level of their logger
* Base time unit is now configurable on MeterRegistry.Config.
* Reworked the histogram API to use a unified builder for both cumulative and normal histograms.

+++ micrometer-core/src/main/java/io/micrometer/core/instrument/stats/hist/Bucket.java
@@ -53,7 +53,7 @@ public class Bucket<T> {
         return this;
     }
 
-    public double getValue() {
-        return value.doubleValue();
+    public long getValue() {
+        return value.longValue();
     }
 }
